kite/requirements1.txt
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
boto3==1.35.71
psycopg2-binary==2.9.10